FIX: add vpoint

+++ Vouchee.Data/Models/DTOs/OrderDTO.cs
@@ -43,10 +43,10 @@ namespace Vouchee.Business.Models.DTOs
         public Guid? promotionId { get; set; }
 
         public string? paymentType { get; set; }
-        public decimal? discountValue { get; set; }
-        public decimal? totalPrice { get; set; }
-        public decimal? discountPrice { get; set; }
-        public decimal? finalPrice { get; set; }
+        public int? discountValue { get; set; }
+        public int? totalPrice { get; set; }
+        public int? discountPrice { get; set; }
+        public int? finalPrice { get; set; }
 
         public string? status { get; set; }
